refactor: split engines, ConfigEntry.runtime_data, async_setup services
- Extract DimmerEngine into engine.py
- Extract CCWCycleEngine into ccw_engine.py
- Add DimmerEngineData TypedDict and DimmerEngineConfigEntry alias
- Replace hass.data[DOMAIN] with entry.runtime_data
- Move service registration to async_setup (one-time, domain-scoped);
  async_setup_entry no longer (un)registers services
- Use entry.async_on_unload for the homeassistant_stop listener
- condition.py reads engines from entry.runtime_data via config_entries

Behavior preserved: all schemas, defaults, and engine logic unchanged.
py_compile clean on all four files.
